third_party/blink/web_tests/html5lib/generated/run-tests1-write-expected.txt

CONSOLE ERROR: Uncaught SyntaxError: Failed to execute 'write' on 'Document': Unexpected token '<'
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select><select>. Please add the missing </select> end tag.
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select><select>. Please add the missing </select> end tag.
../resources/tests1.dat:
30
100

Test 30 of 112 in ../resources/tests1.dat failed. Input:
<select><b><option><select><option></b></select>X
Got:
| <html>
|   <head>
|   <body>
|     <select>
|       <b>
|         <option>
|     <b>
|     <select>
|       <b>
|         <option>
|     "X"
Expected:
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|     <option>
|       "X"

Test 100 of 112 in ../resources/tests1.dat failed. Input:
<select><b><option><select><option></b></select>
Got:
| <html>
|   <head>
|   <body>
|     <select>
|       <b>
|         <option>
|     <b>
|     <select>
|       <b>
|         <option>
Expected:
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|     <option>
